* Now storing waypoints in database table DB_WAYPOINT_TABLE.
NOTE! Need to update config.php with the DB_WAYPOINT_TABLE setting. Also need to set DATABASE_VERSION to 2.99.0 in DB_SETTING_TABLE in database. When site is loaded, the database will be updated and the version number set to 2.99.1. This mechanism will also be used in the future to make database changes.